Tests/LibWeb/Text/expected/lazy-accumulated-visual-context-rebuild.txt

offset width: 110
offset width: 120
offset width: 130
builds after layout-only reads: 0
bounding rect width: 130
builds after transform-sensitive read: 1
